Reduce duplicate code further
